agents/requirements.txt
# Agent dependencies
google-adk>=1.9.0
a2a-sdk>=0.3.0
httpx>=0.27.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
nest-asyncio>=1.6.0
requests>=2.31.0
